Improve mk tools to accept java external libraries
Improve mk tools to add external java libraries to javac classpath to make them
accesible at compilation time.

The java external libraries should be added in the 'lib' section, inside 'requires'
section, in the Cdef file. With the 'z' or '--lib-search' arg in the correspondent
mk tool is possible to provide folders where libraries are searched, allowing to
provide the library name without an absolute path. This way, when the library is
found in the search lib folders, the path is completed to add it to the classpath.

A classpath wildcard is added to the java command to give access to the java external
libraries placed at the lib folder at execution time. The libraries should be bundled
in the Adef file or Cdef file, in the 'bundles' section,  in the lib folder to match
the wildcard or in the lib/ext folder to access them as installed extensions.