Fix bug in compilation when API files with USETYPES are used from Java projects
When in a Java project, an API file with USETYPES tag is included, the compilation
fails because it does not find the import of the USETYPE API file. Also, the code in
C of the USETYPE API file is generated in the project, when we are in a Java project.

In Java this import is not needed because the constants of the USETYPE API file
are resolved and hardcoded in the client implementation, so the imports statements
are removed and the usetype apis generation are ignored in the case of the java
language.